move types around to allow gauntlet import
